+++ python/nvcategory.py
@@ -26,7 +26,7 @@ def to_device(strs):
     return rtn
 
 
-def from_offsets(sbuf, obuf, scount, nbuf=None, ncount=0):
+def from_offsets(sbuf, obuf, scount, nbuf=None, ncount=0, bdevmem=False):
     """
     Create nvcategory object from byte-array of characters encoded in UTF-8.
 
@@ -46,6 +46,8 @@ def from_offsets(sbuf, obuf, scount, nbuf=None, ncount=0):
         null by this bitmask.
     ncount : int, optional
         Optional number of null strings (the default is 0).
+    bdevmem : boolean
+        Default (False) interprets memory pointers as CPU memory.
 
     Examples
     --------
@@ -66,7 +68,8 @@ def from_offsets(sbuf, obuf, scount, nbuf=None, ncount=0):
     ['a', 'e', 'l', 'p'] [0, 3, 3, 2, 1]
 
     """
-    rtn = pyniNVCategory.n_createFromOffsets(sbuf, obuf, scount, nbuf, ncount)
+    rtn = pyniNVCategory.n_createFromOffsets(sbuf, obuf, scount, nbuf, ncount,
+                                             bdevmem)
     if rtn is not None:
         rtn = nvcategory(rtn)
     return rtn