+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - PR #313 Increased regex instruction threshold for stack to 1024
 - PR #302 Add junit test output for py.test
 - PR #331 Use stack-size ranges for different regex instruction counts
+- PR #341 Support device memory in from_offsets
 
 ## Bug Fixes